Replace malloc based memory allocator for CAN message by memory pool
Issue
CAN monitor implemented on PSoC hangs when an intensive traffic
comes to the bus.

Root Cause
CAN implementation uses heap memory to allocate memory for
CAN message objects. The available heap memory is very limited
(128 bytes for PSoC4245AXI-483, no idea if it's possible to extend).
Then malloc stops functioning when several number of RX messages are in.

The malloc is likely to block waiting for any memory gets freed,
but the allocation actually is happening in an interrupt handler
(possibly a bad design, I may want to fix it, too). Freeing the
memory for CAN messages happens always in the application, then
a deadlock situation might be happening.

Fix
Too risky to use malloc in the PSoC platform. The memory allocator
has been replaced by a simple memory pool.